requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
comtypes==1.4.9
filelock==3.17.0
fsspec==2024.12.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
paho-mqtt==2.1.0
pandas==2.2.3
PyAudio==0.2.14
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.2
pywin32==308
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
